feat: APIパスを一つのファイルにまとめた
