+++ backend/src/presentation/routes/mod.rs
@@ -1,3 +1,1 @@
-pub mod auth_routes;
-pub mod todo_routes;
-pub mod routes;
+pub mod api_routes;
